mynewproject/requirements.txt
Django>=4.2.0,<5.0.0
Pillow>=9.0.0
requests>=2.28.0
django-cors-headers>=3.13.0